+++ RayTracer/include/shape/quadric/quadrics.h
@@ -8,16 +8,16 @@ namespace rt
 class Quadric : public Shape
 {
 public:
-	static bool solveQuadraticEq(float* t, float a, float b, float c);
+	static bool solveQuadraticEq(double* t, double a, double b, double c);
 };
 
 struct Sphere : public Quadric
 {
-	float r;
-	glm::vec3 origin;
-	glm::vec3 color;
+	double r;
+	glm::dvec3 origin;
+	glm::dvec3 color;
 
-	Sphere(glm::vec3 origin, float radius, glm::vec3 color, std::shared_ptr<Material> m)
+	Sphere(glm::dvec3 origin, double radius, glm::dvec3 color, std::shared_ptr<Material> m)
 	{
 		this->origin = origin;
 		this->r = radius;
@@ -25,53 +25,53 @@ struct Sphere : public Quadric
 		this->mat = m;
 	}
 
-	glm::vec3 get_normal(glm::vec3 p) const
+	glm::dvec3 get_normal(glm::dvec3 p) const
 	{
 		return glm::normalize(p - origin);
 	}
 
-	float intersect(const Ray &ray, SurfaceInteraction *isect);
+	double intersect(const Ray &ray, SurfaceInteraction *isect);
 
 };
 
 struct Cylinder : public Quadric
 {
-	float height, radius;
-	glm::vec3 pos, dir;
-	glm::mat4 objToWorld;
-	glm::mat4 worldToObj;
-	glm::mat4 tr_worldToObj;
+	double height, radius;
+	glm::dvec3 pos, dir;
+	glm::dmat4 objToWorld;
+	glm::dmat4 worldToObj;
+	glm::dmat4 tr_worldToObj;
 
-	Cylinder(glm::vec3 pos,
-		glm::vec3 dir,
-		float radius,
-		float height,
+	Cylinder(glm::dvec3 pos,
+		glm::dvec3 dir,
+		double radius,
+		double height,
 		std::shared_ptr<Material> mat) :
 		pos(pos), dir(glm::normalize(dir)), radius(radius), height(height)
 	{
 		this->mat = mat;
 
-		glm::vec3 tangent_v = glm::normalize(Plane::getTangentVector(dir));
+		glm::dvec3 tangent_v = glm::normalize(Plane::getTangentVector(dir));
 
 		//objToWorld = glm::lookAt(pos, pos + tangent_v, dir);
 		// transform axis of the cylinder to the axis given by dir
-		objToWorld[0] = glm::vec4(glm::cross(dir, tangent_v), 0.f);
-		objToWorld[1] = glm::vec4(dir, 0.f);
-		objToWorld[2] = glm::vec4(tangent_v, 0.f);
-		objToWorld[3] = glm::vec4(pos, 1.f);
+		objToWorld[0] = glm::dvec4(glm::cross(dir, tangent_v), 0.f);
+		objToWorld[1] = glm::dvec4(dir, 0.f);
+		objToWorld[2] = glm::dvec4(tangent_v, 0.f);
+		objToWorld[3] = glm::dvec4(pos, 1.f);
 
 		worldToObj = glm::inverse(objToWorld);
 		tr_worldToObj = glm::transpose(worldToObj);
 	}
 
-	float intersect(const Ray &ray, SurfaceInteraction *isect);
+	double intersect(const Ray &ray, SurfaceInteraction *isect);
 
-	glm::vec3 get_normal(glm::vec3 p, int hit_cnt) const
+	glm::dvec3 get_normal(glm::dvec3 p, int hit_cnt) const
 	{
 		if (hit_cnt == 2)
-			return glm::normalize(tr_worldToObj * glm::vec4(p.x, 0.f, p.z, 0.f));
+			return glm::normalize(tr_worldToObj * glm::dvec4(p.x, 0.f, p.z, 0.f));
 		else
-			return -glm::normalize((tr_worldToObj * glm::vec4(p.x, 0.f, p.z, 0.f)));
+			return -glm::normalize((tr_worldToObj * glm::dvec4(p.x, 0.f, p.z, 0.f)));
 	}
 };
